callfun/requirements_enhanced.txt
# Enhanced AI Grocery Store Assistant Requirements
# Core dependencies for conversation state management

# HTTP and API
requests>=2.31.0
flask>=2.3.0
flask-cors>=4.0.0

# Database
pymongo>=4.5.0
certifi>=2023.7.22

# AI and LLM
google-generativeai>=0.3.0

# Voice processing
SpeechRecognition>=3.10.0
pyttsx3>=2.90

# Data processing
python-dotenv>=1.0.0

# Threading and concurrency
threading2>=0.3.1

# Date and time
python-dateutil>=2.8.2

# JSON and data handling
jsonschema>=4.19.0

# Logging and monitoring
structlog>=23.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Production deployment
gunicorn>=21.2.0
uvicorn>=0.23.0

# Optional: For enhanced voice processing
# pyaudio>=0.2.11  # Uncomment if using PyAudio for better audio handling
# webrtcvad>=2.0.10  # Uncomment for voice activity detection 